# docs/openapi-spec.yaml
openapi: 3.0.0
info:
  title: task-management
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/tasks/{id}:
    get:
      tags:
        - v1 > tasks
      summary: get by id
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '6'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249904825473'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '484'
            ETag:
              schema:
                type: string
                example: W/"1e4-78qp08FlF54f6bDjylRjHiMjRMY"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 03:56:38 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 6
                  title: task1
                  description: this is task 1
                  dueDate: '2024-09-01T04:00:00.000Z'
                  priority: null
                  storyPoint: null
                  status: OPEN
                  createdById: 1
                  createdAt: '2024-08-31T10:49:55.728Z'
                  updatedAt: '2024-08-31T10:49:55.728Z'
                  deletedAt: null
                  labels:
                    - label:
                        id: 27
                        name: dev
                    - label:
                        id: 28
                        name: backend
                  assignees:
                    - user:
                        id: 1
                        username: user1
                    - user:
                        id: 2
                        username: user2
                  createdBy:
                    id: 1
                    username: user1
    put:
      tags:
        - v1 > tasks
      summary: replace one task
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                title: task5
                description: this is task 1
                dueDate: '2024-09-01 12:00:00.000'
                createdBy: 1
                labels:
                  - new
                assignees:
                  - 3
                  - 4
                storyPoint: 1
                status: OPEN
                priority: null
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '5'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249912835182'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '480'
            ETag:
              schema:
                type: string
                example: W/"1e0-WKSUZ0s33JnkcJiU/0dVbdv96CA"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 04:09:59 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 5
                  title: task5
                  description: this is task 1
                  dueDate: '2024-09-01T04:00:00.000Z'
                  priority: null
                  storyPoint: null
                  status: OPEN
                  createdById: 1
                  createdAt: '2024-08-31T10:49:21.701Z'
                  updatedAt: '2024-09-01T04:09:59.941Z'
                  deletedAt: null
                  labels:
                    - label:
                        id: 27
                        name: dev
                    - label:
                        id: 35
                        name: new
                  assignees:
                    - user:
                        id: 1
                        username: user1
                    - user:
                        id: 3
                        username: user3
                  createdBy:
                    id: 1
                    username: user1
    patch:
      tags:
        - v1 > tasks
      summary: update one
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                title: task5
                labels:
                  - good
                assignees:
                  - 4
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '5'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17250047746459'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '440'
            ETag:
              schema:
                type: string
                example: W/"1b8-sx9AuYgnjWmStrcRyvmYRNd9BbQ"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 07:54:52 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 5
                  title: task5
                  description: this is task 1
                  dueDate: '2024-09-01T04:00:00.000Z'
                  priority: null
                  storyPoint: 1
                  status: OPEN
                  createdById: 1
                  createdAt: '2024-08-31T10:49:21.701Z'
                  updatedAt: '2024-09-01T07:54:52.587Z'
                  deletedAt: null
                  labels:
                    - label:
                        id: 35
                        name: new
                    - label:
                        id: 44
                        name: wow
                  assignees:
                    - user:
                        id: 3
                        username: user3
                  createdBy:
                    id: 1
                    username: user1
    delete:
      tags:
        - v1 > tasks
      summary: delete a task
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '1'
      responses:
        '204':
          description: No Content
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249913028118'
            ETag:
              schema:
                type: string
                example: W/"e-hDOZGZ11/MTPwZq/A8sGFlQzkPc"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 04:10:23 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            text/plain:
              schema:
                type: string
              example: null
  /v1/tasks:
    get:
      tags:
        - v1 > tasks
      summary: query
      parameters:
        - name: filters[status]
          in: query
          schema:
            type: string
          example: OPEN
        - name: sort[0]
          in: query
          schema:
            type: string
          example: createdAt:asc
        - name: filters[title]
          in: query
          schema:
            type: string
          example: TASk5
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249907846710'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '543'
            ETag:
              schema:
                type: string
                example: W/"21f-auRCDJcGyFxNz37RZfNQmaaVQic"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 04:01:41 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  - id: 5
                    title: task5
                    description: this is task 1
                    dueDate: '2024-09-01T04:00:00.000Z'
                    priority: null
                    storyPoint: null
                    status: OPEN
                    createdById: 1
                    createdAt: '2024-08-31T10:49:21.701Z'
                    updatedAt: '2024-08-31T12:30:09.870Z'
                    deletedAt: null
                    labels:
                      - label:
                          id: 27
                          name: dev
                      - label:
                          id: 35
                          name: new
                    assignees:
                      - user:
                          id: 1
                          username: user1
                      - user:
                          id: 3
                          username: user3
                    createdBy:
                      id: 1
                      username: user1
                pagination:
                  total: 2
                  pageCount: 2
                  page: 1
                  pageSize: 1
    post:
      tags:
        - v1 > tasks
      summary: create new task
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                title: task99
                description: this is task 99
                dueDate: '2024-09-01 12:00:00.000'
                createdBy: 1
                labels:
                  - dev
                assignees:
                  - 4
      responses:
        '201':
          description: Created
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249912362039'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '412'
            ETag:
              schema:
                type: string
                example: W/"19c-cDK/fPojR3oXMujk+/JmFK/w3fQ"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 04:09:12 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 8
                  title: task99
                  description: this is task 99
                  dueDate: '2024-09-01T04:00:00.000Z'
                  priority: null
                  storyPoint: null
                  status: OPEN
                  createdById: 1
                  createdAt: '2024-09-01T04:09:12.616Z'
                  updatedAt: '2024-09-01T04:09:12.616Z'
                  deletedAt: null
                  labels:
                    - label:
                        id: 27
                        name: dev
                  assignees:
                    - user:
                        id: 4
                        username: user4
                  createdBy:
                    id: 1
                    username: user1